backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
black==23.10.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
databases==0.8.0
dnspython==2.4.2
docopt==0.6.2
email-validator==2.1.0
exceptiongroup==1.1.3
fastapi==0.104.0
FastAPI-SQLAlchemy==0.2.1
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
mysql-connector==2.2.9
orjson==3.9.9
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.11.0
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
thingspeak==1.0.0
tomli==2.0.1
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.0.7
uvicorn==0.23.2
uvloop==0.18.0
watchfiles==0.21.0
websockets==12.0
